refactor modaps download code out
